+++ demo/forms.py
@@ -5,10 +5,19 @@ from django import forms
 from django.core.urlresolvers import reverse_lazy
 
 from agnocomplete import fields, widgets
-from agnocomplete.forms import UserContextForm
+from agnocomplete.forms import UserContextFormMixin
 
-from demo.autocomplete import AutocompleteColor, AutocompletePerson
-from demo.autocomplete import HiddenAutocomplete
+from .autocomplete import (
+    AutocompleteColor,
+    AutocompleteColorShort,
+    AutocompletePerson,
+    AutocompletePersonShort,
+    HiddenAutocomplete,
+    AutocompleteTag,
+    AutocompleteContextTag,
+)
+from .models import PersonTag, PersonContextTag
+from .fields import ModelMultipleDomainField
 
 
 class SearchForm(forms.Form):
@@ -27,7 +36,7 @@ class SearchFormTextInput(forms.Form):
         AutocompletePerson, widget=widgets.AgnocompleteTextInput)
 
 
-class SearchContextForm(UserContextForm):
+class SearchContextForm(UserContextFormMixin, forms.Form):
     search_person = fields.AgnocompleteModelField('AutocompletePersonDomain')
 
 
@@ -35,3 +44,47 @@ class SearchCustom(forms.Form):
     search_color = fields.AgnocompleteField(
         HiddenAutocomplete(url=reverse_lazy('hidden-autocomplete')),
     )
+
+
+class SearchColorMulti(forms.Form):
+    search_multi_color = fields.AgnocompleteMultipleField(
+        AutocompleteColorShort)
+    search_multi_color_create = fields.AgnocompleteMultipleField(
+        AutocompleteColorShort,
+        create=True,
+    )
+
+
+class PersonTagForm(forms.Form):
+    person = fields.AgnocompleteModelField(AutocompletePersonShort)
+    tags = fields.AgnocompleteModelMultipleField(AutocompleteTag)
+
+
+class PersonTagModelForm(forms.ModelForm):
+    person = fields.AgnocompleteModelField(AutocompletePersonShort)
+    tags = fields.AgnocompleteModelMultipleField(AutocompleteTag)
+
+    class Meta:
+        model = PersonTag
+        fields = '__all__'
+
+
+class PersonTagModelFormWithCreate(PersonTagModelForm):
+    tags = fields.AgnocompleteModelMultipleField(
+        AutocompleteTag,
+        create_field="name",
+        required=False
+    )
+
+
+class PersonContextTagModelForm(UserContextFormMixin, forms.ModelForm):
+    person = fields.AgnocompleteModelField(AutocompletePersonShort)
+    tags = ModelMultipleDomainField(
+        AutocompleteContextTag,
+        create_field="name",
+        required=False
+    )
+
+    class Meta:
+        model = PersonContextTag
+        fields = '__all__'